# docs/swagger.yaml
basePath: /
consumes:
- application/json
definitions:
  wazmeow_internal_http_dto.ConnectSessionResponse:
    description: Resposta da operação de conexão de sessão
    properties:
      message:
        example: QR Code gerado. Escaneie com seu WhatsApp.
        type: string
      needs_auth:
        example: true
        type: boolean
      qr_code:
        example: data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAA...
        type: string
      session:
        $ref: '#/definitions/wazmeow_internal_http_dto.SessionResponse'
    type: object
...
  wazmeow_internal_http_dto.DeleteSessionResponse:
    properties:
      message:
        type: string
      session_id:
        type: string
    type: object
  wazmeow_internal_http_dto.DisconnectSessionResponse:
    properties:
      message:
        type: string
      session:
        $ref: '#/definitions/wazmeow_internal_http_dto.SessionResponse'
    type: object
  wazmeow_internal_http_dto.ErrorResponse:
    description: Resposta de erro padrão da API
    properties:
      code:
        example: INTERNAL_ERROR
        type: string
      context:
        additionalProperties: {}
        type: object
      details:
        example: Detalhes técnicos do erro
        type: string
      error:
        example: Erro interno do servidor
        type: string
      status:
        example: error
        type: string
      success:
        example: false
        type: boolean
      timestamp:
        example: "2024-01-01T12:00:00Z"
        type: string
    type: object
...
  wazmeow_internal_http_dto.PairPhoneRequest:
    description: Dados para emparelhamento com número de telefone
    properties:
      phone_number:
        example: "5511999999999"
        type: string
    required:
    - phone_number
    type: object
...
  wazmeow_internal_http_dto.SuccessResponse:
    description: Resposta de sucesso padrão da API
    properties:
      data: {}
      message:
        example: Operação realizada com sucesso
        type: string
      status:
        example: success
        type: string
      success:
        example: true
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@wazmeow.com
    name: WazMeow API Support
    url: https://github.com/wazmeow/wazmeow
  description: 'Autenticação básica HTTP (username:password). Configure AUTH_ENABLED=true
    e AUTH_TYPE=basic no .env. Exemplo: Authorization: Basic dXNlcjpwYXNz'
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://github.com/wazmeow/wazmeow/blob/main/LICENSE
  title: WazMeow API
  version: 1.0.0
paths:
...
  /sessions/{id}:
    delete:
      description: Deleta uma sessão WhatsApp específica por ID ou nome. Sempre força
        a deleção mesmo se conectada
      parameters:
      - description: ID da sessão (UUID) ou nome da sessão
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Sessão deletada
          schema:
            allOf:
            - $ref: '#/definitions/wazmeow_internal_http_dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/wazmeow_internal_http_dto.DeleteSessionResponse'
              type: object
        "400":
          description: Identificador da sessão inválido
          schema:
            $ref: '#/definitions/wazmeow_internal_http_dto.ErrorResponse'
        "404":
          description: Sessão não encontrada
          schema:
            $ref: '#/definitions/wazmeow_internal_http_dto.ErrorResponse'
        "500":
          description: Erro interno
          schema:
            $ref: '#/definitions/wazmeow_internal_http_dto.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Deletar sessão WhatsApp
      tags:
      - Sessions
  /sessions/{id}/connect:
    post:
      consumes:
      - application/json
      description: |-
        Inicia o processo de conexão de uma sessão WhatsApp. Se a sessão não estiver autenticada, gera um QR Code para escaneamento.

        **Fluxo de conexão:**
        1. Sessão não autenticada: Retorna QR Code para escaneamento
        2. Sessão autenticada: Conecta diretamente ao WhatsApp
        3. Sessão já conectada: Retorna erro 409

        **Identificadores aceitos:**
        - UUID da sessão: `4ee6195b-6a0f-4c85-a4ee-673ee15f14c8`
        - Nome da sessão: `minha-sessao`
      parameters:
      - description: ID da sessão (UUID) ou nome da sessão
        example: '"minha-sessao"'
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Processo de conexão iniciado (QR Code gerado ou sessão conectada)
          schema:
            allOf:
            - $ref: '#/definitions/wazmeow_internal_http_dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/wazmeow_internal_http_dto.ConnectSessionResponse'
              type: object
        "400":
          description: Identificador da sessão inválido ou malformado
          schema:
            $ref: '#/definitions/wazmeow_internal_http_dto.ErrorResponse'
        "404":
          description: Sessão não encontrada com o identificador fornecido
          schema:
            $ref: '#/definitions/wazmeow_internal_http_dto.ErrorResponse'
        "409":
          description: Sessão já está conectada
          schema:
            $ref: '#/definitions/wazmeow_internal_http_dto.ErrorResponse'
        "500":
          description: Erro interno do servidor ou falha na conexão WhatsApp
          schema:
            $ref: '#/definitions/wazmeow_internal_http_dto.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Conectar sessão WhatsApp
      tags:
      - Sessions
  /sessions/{id}/info:
    get:
      consumes:
      - application/json
      description: Retorna as informações detalhadas de uma sessão específica por
        ID ou nome, incluindo status completo
      parameters:
      - description: ID da sessão (UUID) ou nome da sessão
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Detalhes da sessão
          schema:
            allOf:
            - $ref: '#/definitions/wazmeow_internal_http_dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/wazmeow_internal_http_dto.SessionResponse'
              type: object
        "400":
          description: Identificador da sessão inválido
          schema:
            $ref: '#/definitions/wazmeow_internal_http_dto.ErrorResponse'
        "404":
          description: Sessão não encontrada
          schema:
            $ref: '#/definitions/wazmeow_internal_http_dto.ErrorResponse'
        "500":
          description: Erro interno
          schema:
            $ref: '#/definitions/wazmeow_internal_http_dto.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Obter detalhes da sessão
      tags:
      - Sessions
  /sessions/{id}/logout:
    post:
      consumes:
      - application/json
      description: Desconecta a sessão do WhatsApp, encerrando a comunicação
      parameters:
      - description: ID da sessão (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Sessão desconectada
          schema:
            allOf:
            - $ref: '#/definitions/wazmeow_internal_http_dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/wazmeow_internal_http_dto.DisconnectSessionResponse'
              type: object
        "400":
          description: ID da sessão inválido
          schema:
            $ref: '#/definitions/wazmeow_internal_http_dto.ErrorResponse'
        "404":
          description: Sessão não encontrada
          schema:
            $ref: '#/definitions/wazmeow_internal_http_dto.ErrorResponse'
        "409":
          description: Sessão já desconectada
          schema:
            $ref: '#/definitions/wazmeow_internal_http_dto.ErrorResponse'
        "500":
          description: Erro interno
          schema:
            $ref: '#/definitions/wazmeow_internal_http_dto.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Desconectar sessão (logout)
      tags:
      - Sessions
  /sessions/{id}/pairphone:
    post:
      consumes:
      - application/json
      description: Emparelha um telefone com a sessão WhatsApp por ID ou nome
      parameters:
      - description: ID da sessão (UUID) ou nome da sessão
        in: path
        name: id
        required: true
        type: string
      - description: Dados do telefone
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/wazmeow_internal_http_dto.PairPhoneRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Telefone emparelhado
          schema:
            allOf:
            - $ref: '#/definitions/wazmeow_internal_http_dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/wazmeow_internal_http_dto.PairPhoneResponse'
              type: object
        "400":
          description: Dados inválidos
          schema:
            $ref: '#/definitions/wazmeow_internal_http_dto.ErrorResponse'
        "404":
          description: Sessão não encontrada
          schema:
            $ref: '#/definitions/wazmeow_internal_http_dto.ErrorResponse'
        "500":
          description: Erro interno
          schema:
            $ref: '#/definitions/wazmeow_internal_http_dto.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Emparelhar telefone com sessão
      tags:
      - Sessions
  /sessions/{id}/proxy/set:
    post:
      consumes:
      - application/json
      description: |-
        Configura ou atualiza a configuração de proxy para uma sessão existente. O proxy será usado para todas as conexões WhatsApp desta sessão.

        **Tipos de proxy suportados:**
        - HTTP: Proxy HTTP/HTTPS padrão
        - SOCKS5: Proxy SOCKS5 com suporte a autenticação

        **Exemplos de configuração:**
        - Proxy HTTP: `{"proxy_host": "78.24.204.134", "proxy_port": 62122, "proxy_type": "http"}`
        - Proxy com autenticação: `{"proxy_host": "78.24.204.134", "proxy_port": 62122, "proxy_type": "http", "username": "user", "password": "pass"}`
        - Proxy SOCKS5: `{"proxy_host": "78.24.204.134", "proxy_port": 62123, "proxy_type": "socks5", "username": "user", "password": "pass"}`
      parameters:
      - description: ID da sessão (UUID) ou nome da sessão
        example: '"minha-sessao"'
        in: path
        name: id
        required: true
        type: string
      - description: Configuração do proxy
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/wazmeow_internal_http_dto.ProxySetRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Proxy configurado com sucesso
          schema:
            allOf:
            - $ref: '#/definitions/wazmeow_internal_http_dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/wazmeow_internal_http_dto.ProxySetResponse'
              type: object
        "400":
          description: Dados de proxy inválidos (host, porta, tipo, etc.)
          schema:
            $ref: '#/definitions/wazmeow_internal_http_dto.ErrorResponse'
        "404":
          description: Sessão não encontrada
          schema:
            $ref: '#/definitions/wazmeow_internal_http_dto.ErrorResponse'
        "422":
          description: 'Configuração de proxy inválida (ex: username sem password)'
          schema:
            $ref: '#/definitions/wazmeow_internal_http_dto.ErrorResponse'
        "500":
          description: Erro interno do servidor
          schema:
            $ref: '#/definitions/wazmeow_internal_http_dto.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Configurar proxy para sessão
      tags:
      - Sessions
...
produces:
- application/json
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    description: 'API Key para autenticação. Configure AUTH_ENABLED=true e AUTH_TYPE=api_key
      no .env. Exemplo: X-API-Key: sua-chave-api-aqui'
    in: header
    name: X-API-Key
    type: apiKey
  BasicAuth:
    type: basic
swagger: "2.0"
...